Move helper functions into separate module cli_helpers.py to avoid circular imports
