+++ qcrbox/qcrbox/cli/helpers/__init__.py
@@ -2,11 +2,10 @@
 
 import click
 
+from .cli_helpers import add_cli_option_enable_disable_components, add_verbose_option
 from .docker_project import DockerProject
 from .doit_helpers import make_task, run_tasks
 from .qcrbox_helpers import (
-    add_cli_option_enable_disable_components,
-    add_verbose_option,
     get_mkdocs_config_file_path,
     get_repo_root,
 )
